feat: Integrate serverless-http and update server start logic for local development and deployement on vercel
