+++ backend/api/index.ts
@@ -1,0 +1,4 @@
+import serverless from "serverless-http";
+import app from "../server.ts";
+
+export default serverless(app);
